test_requirements.txt
coverage
mock
pytest
python-dotenv
waiting
